requirements.txt
discord.py
python-dotenv
gtts
pyNaCl